feat(decox-react): mapToProps -> propsMapper
mapToProps: (store, wrapperProps) -> props

If you define a function in it, the function is re-defined every time
`mapToProps` is called. This causes unnecessary re-rendering
since the `shallowEqual` with the preceding props will always be false.

propsMapper: store -> wrapperProps -> props

It defines a function that defines `mapToProps` so you can predefine
function you want to use in its closure and use it without re-defining
it.